backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
PyJWT[crypto]==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
reportlab==4.1.0
openpyxl==3.1.2
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
